+++ metrics/tarantool.lua
@@ -22,7 +22,6 @@ local default_metrics = {
     luajit              = require('metrics.tarantool.luajit'),
     cartridge_issues    = require('metrics.cartridge.issues'),
     cartridge_failover  = require('metrics.cartridge.failover'),
-    cartridge_config    = require('metrics.cartridge.config'),
     clock               = require('metrics.tarantool.clock'),
     event_loop          = require('metrics.tarantool.event_loop'),
     config              = require('metrics.tarantool.config'),