+++ CHANGELOG.md
@@ -5,39 +5,45 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
+
 - `tnt_cartridge_config_applied` metric.
 - New optional ``label_keys`` parameter for ``counter()`` and ``gauge()`` metrics.
 
 ## [1.3.1] - 2025-02-24
 
 ## [1.3.0] - 2025-02-24
+
 ### Added
+
 - New metrics category in config: ``cpu_extended``.
 
 ### Fixed
+
 - Use box.info.ro instead of box.cfg.read_only in replication metrics.
 
 ## [1.2.0] - 2024-07-09
+
 ### Added
+
 - New Tarantool 3 metrics:
   - tnt_config_alerts
   - tnt_config_status
 
 ## [1.1.0] - 2024-05-17
+
 ### Added
+
 - `metrics.cfg{}` `"all"` metasection for array `include` and `exclude`
   (`metrics.cfg{include={'all'}}` can be used instead of `metrics.cfg{include='all'}`,
   `metrics.cfg{exclude={'all'}}` can be used instead of `metrics.cfg{include='none'}`)
-
 - `tnt_election_leader_idle` metric.
-
 - Histogram now logs a warning if `observe` is called with `cdata` value.
 
 ### Changed
 
 - Inconsistent metrics descrtiptions for memtx metrics.
-
 - New memory metrics:
   - tnt_memtx_tuples_data_total
   - tnt_memtx_tuples_data_read_view
@@ -51,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Using `cdata` values with `histogram:observe`.
 
 ## [1.0.0] - 2023-05-22
+
 ### Changed
 
 - Don't export self replication status.
@@ -61,7 +68,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (**incompatible change**).
 
 ## [0.17.0] - 2023-03-23
+
 ### Added
+
 - `metrics.cfg{}` -- a single entrypoint to setup the module:
   - `include` and `exclude` options with the same effect as in
     `enable_default_metrics(include, exclude)` (but its deprecated
@@ -73,6 +82,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Versioning support through `require('metrics')._VERSION`
 
 ### Changed
+
 - Setup cartridge hotreload inside the role
 - Extend `enable_default_metrics()` API:
   - `'all'` and `'none'` options for `include` argument,
@@ -82,12 +92,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Override built-in metrics, if installed
 
 ### Deprecated
+
 - Passing nonexistent metrics to `enable_default_metrics()`
 - Using `{}` as `include` in `enable_default_metrics()`
   to enable all metrics
 - Versioning support through `require('metrics').VERSION`
 
 ## [0.16.0] - 2023-01-27
+
 ### Added
 
 - Handle to clear psutils metrics
@@ -109,6 +121,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deprecated metrics from previous releases
 
 ## [0.15.1] - 2022-09-20
+
 ### Added
 
  - `memtx` MVCC memory monitoring
@@ -120,6 +133,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix queries in quantile
 
 ## [0.15.0] - 2022-08-09
+
 ### Fixed
 
 - Clean info about spaces and indexes after their drop
@@ -173,6 +187,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deprecated metrics from previous releases
 
 ## [0.14.0] - 2022-06-28
+
 ### Fixed
 
 - Float numbers in Graphite exporter
@@ -195,6 +210,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `enable_cartridge_metrics` function
 
 ## [0.13.0] - 2022-03-23
+
 ### Fixed
 
 - Don't reset collectors when Cartridge roles hot reload
@@ -249,44 +265,57 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `tnt_replication_replica_<id>_lsn`
 
 ## [0.12.0] - 2021-11-18
+
 ### Changed
+
 - Update `http` dependency to 1.1.1
 
 ### Fixed
+
 - Cast number64 to json number in json export plugin [#321](https://github.com/tarantool/metrics/issues/321)
 
 ### Deprecated
+
 - HTTP middleware v2
 
 ## [0.11.0] - 2021-09-23
+
 ### Added
+
 - collector's method `remove` to clear observations with given label pairs [#263](https://github.com/tarantool/metrics/issues/263)
 - `counter:reset()` method [#260](https://github.com/tarantool/metrics/issues/260)
 - `tnt_read_only` metric [#275](https://github.com/tarantool/metrics/issues/275)
 
 ### Removed
+
 - Average collector
 
 ### Fixed
+
 - Throw an error when http_middelware is processing a wrong handler [#199](https://github.com/tarantool/metrics/issues/199)
 - cartridge issues metric fails before cartridge.cfg() call [#298](https://github.com/tarantool/metrics/issues/298)
 
 ### Changed
+
 - quantile metric is NAN if no samples provided for an age [#303](https://github.com/tarantool/metrics/issues/303)
 
 ## [0.10.0] - 2021-08-03
+
 ### Changed
+
 - metrics registry refactoring to search with `O(1)` [#188](https://github.com/tarantool/metrics/issues/188)
 - `ipairs` instead of `pairs` while iteration in `histogram` [#196](https://github.com/tarantool/metrics/issues/196)
 - `set_export` function provide default metrics config to make role reloadable [#248](https://github.com/tarantool/metrics/issues/248)
 - metrics registry refactoring to add and remove callbacks with `O(1)` [#276](https://github.com/tarantool/metrics/issues/276)
 
 ### Fixed
+
 - be gentle to http routes, don't leave gaps in the array
   [#246](https://github.com/tarantool/metrics/issues/246)
 - allow to create summary without observations [#265](https://github.com/tarantool/metrics/issues/265)
 
 ### Added
+
 - `tnt_clock_delta` metric to compute clock difference on instances [#232](https://github.com/tarantool/metrics/issues/232)
 - set custom global labels in config and with `set_labels` function [#259](https://github.com/tarantool/metrics/issues/259)
 - allow to include and exclude default metrics in config and in `enable_default_metrics` function
@@ -297,38 +326,52 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `enable_cartridge_metrics` function
 
 ## [0.9.0] - 2021-05-28
+
 ### Fixed
+
 - cartridge metrics role fails to start without http [#225](https://github.com/tarantool/metrics/issues/225)
 - quantile overflow after `fiber.yield()` [#235](https://github.com/tarantool/metrics/issues/235)
 - role and module hot reload [#227](https://github.com/tarantool/metrics/issues/227), [#228](https://github.com/tarantool/metrics/issues/228)
 
 ### Changed
+
 - `tnt_cartridge_issues` gathers only local issues [#211](https://github.com/tarantool/metrics/issues/211)
 
 ### Added
+
 - Age buckets in `summary`
 
 ## [0.8.0] - 2021-04-13
+
 ### Added
+
 - New default metrics: cpu_user_time, cpu_system_time
 - Vinyl metrics
 
 ## [0.7.1] - 2021-03-18
+
 ### Added
+
 - zone label support for Tarantool Cartridge >= '2.4.0'
 - rpm packaging for CentOS 8, Fedora 30, 31, 32
 
 ## [0.7.0] - 2021-02-09
+
 ### Added
+
 - instance health check plugin
 
 ## [0.6.1] - 2021-01-20
+
 ### Fixed
+
 - package reloading works for `metrics.quantile`
 - instance_name in alias label if no alias present
 
 ## [0.6.0] - 2020-12-01
+
 ### Fixed
+
 - metrics.clear() disables default metrics
 - cartridge role is permanent
 - cartridge role configuration without clusterwide config
@@ -338,47 +381,64 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - graphite plugin allows empty prefix
 
 ### Added
+
 - Luajit platform metrics
 - `enable_cartridge_metrics` function
 - Cartridge issues gauge
 
 ### Changed
+
 - CI on Github Actions
 
 ## [0.5.0] - 2020-09-18
+
 ### Added
+
 - Summary collector
 
 ### Deprecated
+
 - Average collector
 
 ## [0.4.0] - 2020-07-14
+
 ### Added
+
 - New default metrics: cpu_total, cpu_thread
 - histogram:observe_latency for measure latency of function call with example
 
 ## [0.3.0] - 2020-06-11
+
 ### Added
+
 - Role for [tarantool/cartridge](https://github.com/tarantool/cartridge)
 - Documentaion and examples on [tarantool/http server](https://github.com/tarantool/http) middleware
 
 ### Fixed
+
 - Throw exception when `http_middleware.build_default_collector` is called with same name
 - Attempt to index non-existent master vclock on a replica after the death of the master
 
 ## [0.2.0] - 2020-05-07
+
 ### Added
 - [tarantool/http server](https://github.com/tarantool/http) middleware to collect http server metrics
+
 ### Fixed
+
 - `collect` failure for vinyl metrics in strict mode
 - prometheus exporter: render 0 for 0ULL value instead of +Inf
 - Travis CI build failures
+
 ### Changed
+
 - Renamed `info_vclock_{ID}` metric to `info_vclock`, moved `{ID}` to tags
 - Renamed `stat_op_{OP_TYPE}_total` metric to `stat_op_total`, moved `{OP_TYPE}` to tags
 - Renamed `stat_op_{OP_TYPE}_rps` metric to `stat_op_rps`, moved `{OP_TYPE}` to tags
 - Renamed `space_index_{IDX_NAME}_bsize` to `space_index_bsize`, moved `{IDX_NAME}` to tags
 
 ## [0.1.8] - 2020-01-15
+
 ### Added
+
 - Ability to set labels globally for each metric