+++ CHANGELOG.md
@@ -8,11 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `tnt_cartridge_config_checksum` metric.
-
 ### Changed
 
 ### Fixed
+
+# [1.5.0] - 2025-08-13
+
+### Added
+
+- `tnt_cartridge_config_checksum` metric.
 
 ### Removed